feat(starship): build of starship

+++ dot_bootstrap/build.sh
@@ -13,15 +13,20 @@ rust () {
 }
 
 alacritty () {
-	source /root/.cargo/env
-	cargo install alacritty
-	mv /root/.cargo/bin/alacritty /usr/local/bin/alacritty
+	if [[ ! -f "/usr/local/bin/alacritty" ]]; then
+		source /root/.cargo/env
+		cargo install alacritty
+		mv /root/.cargo/bin/alacritty /usr/local/bin/alacritty
+	fi
 }
 
 fzf () {
 	git clone --depth 1 https://github.com/junegunn/fzf.git /root/.fzf
 	/root/.fzf/install --bin
-	mv /usr/bin/fzf /usr/local/bin/fzf
+}
+
+starship() {
+	curl -sS https://starship.rs/install.sh | sh -s -- -y
 }
 
 $*
